+++ log/ratelimit.go
@@ -17,21 +17,27 @@ type RateLimitedLogger struct {
 	next    Interface
 	limiter *rate.Limiter
 
-	discardedLogLinesCounter *prometheus.CounterVec
+	discardedInfoLogLinesCounter  prometheus.Counter
+	discardedDebugLogLinesCounter prometheus.Counter
+	discardedWarnLogLinesCounter  prometheus.Counter
+	discardedErrorLogLinesCounter prometheus.Counter
 }
 
 // NewRateLimitedLogger returns a logger.Interface that is limited to the given number of logs per second,
 // with the given burst size.
 func NewRateLimitedLogger(logger Interface, logsPerSecond rate.Limit, burstSize int, reg prometheus.Registerer) Interface {
 	discardedLogLinesCounter := promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
-		Name: "rate_limit_logger_discarded_log_lines_total",
+		Name: "logger_rate_limit_discarded_log_lines_total",
 		Help: "Total number of discarded log lines per level.",
 	}, []string{"level"})
 
 	return &RateLimitedLogger{
-		next:                     logger,
-		limiter:                  rate.NewLimiter(logsPerSecond, burstSize),
-		discardedLogLinesCounter: discardedLogLinesCounter,
+		next:                          logger,
+		limiter:                       rate.NewLimiter(logsPerSecond, burstSize),
+		discardedInfoLogLinesCounter:  discardedLogLinesCounter.WithLabelValues(infoLevel),
+		discardedDebugLogLinesCounter: discardedLogLinesCounter.WithLabelValues(debugLevel),
+		discardedWarnLogLinesCounter:  discardedLogLinesCounter.WithLabelValues(warnLevel),
+		discardedErrorLogLinesCounter: discardedLogLinesCounter.WithLabelValues(errorLevel),
 	}
 }
 
@@ -39,7 +45,7 @@ func (l *RateLimitedLogger) Debugf(format string, args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Debugf(format, args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(debugLevel).Inc()
+		l.discardedDebugLogLinesCounter.Inc()
 	}
 }
 
@@ -47,7 +53,7 @@ func (l *RateLimitedLogger) Debugln(args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Debugln(args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(debugLevel).Inc()
+		l.discardedDebugLogLinesCounter.Inc()
 	}
 }
 
@@ -55,7 +61,7 @@ func (l *RateLimitedLogger) Infof(format string, args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Infof(format, args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(infoLevel).Inc()
+		l.discardedInfoLogLinesCounter.Inc()
 	}
 }
 
@@ -63,7 +69,7 @@ func (l *RateLimitedLogger) Infoln(args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Infoln(args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(infoLevel).Inc()
+		l.discardedInfoLogLinesCounter.Inc()
 	}
 }
 
@@ -71,7 +77,7 @@ func (l *RateLimitedLogger) Errorf(format string, args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Errorf(format, args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(errorLevel).Inc()
+		l.discardedErrorLogLinesCounter.Inc()
 	}
 }
 
@@ -79,7 +85,7 @@ func (l *RateLimitedLogger) Errorln(args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Errorln(args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(errorLevel).Inc()
+		l.discardedErrorLogLinesCounter.Inc()
 	}
 }
 
@@ -87,7 +93,7 @@ func (l *RateLimitedLogger) Warnf(format string, args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Warnf(format, args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(warnLevel).Inc()
+		l.discardedWarnLogLinesCounter.Inc()
 	}
 }
 
@@ -95,22 +101,28 @@ func (l *RateLimitedLogger) Warnln(args ...interface{}) {
 	if l.limiter.Allow() {
 		l.next.Warnln(args...)
 	} else {
-		l.discardedLogLinesCounter.WithLabelValues(warnLevel).Inc()
+		l.discardedWarnLogLinesCounter.Inc()
 	}
 }
 
 func (l *RateLimitedLogger) WithField(key string, value interface{}) Interface {
 	return &RateLimitedLogger{
-		next:                     l.next.WithField(key, value),
-		limiter:                  l.limiter,
-		discardedLogLinesCounter: l.discardedLogLinesCounter,
+		next:                          l.next.WithField(key, value),
+		limiter:                       l.limiter,
+		discardedInfoLogLinesCounter:  l.discardedInfoLogLinesCounter,
+		discardedDebugLogLinesCounter: l.discardedDebugLogLinesCounter,
+		discardedWarnLogLinesCounter:  l.discardedWarnLogLinesCounter,
+		discardedErrorLogLinesCounter: l.discardedErrorLogLinesCounter,
 	}
 }
 
 func (l *RateLimitedLogger) WithFields(f Fields) Interface {
 	return &RateLimitedLogger{
-		next:                     l.next.WithFields(f),
-		limiter:                  l.limiter,
-		discardedLogLinesCounter: l.discardedLogLinesCounter,
+		next:                          l.next.WithFields(f),
+		limiter:                       l.limiter,
+		discardedInfoLogLinesCounter:  l.discardedInfoLogLinesCounter,
+		discardedDebugLogLinesCounter: l.discardedDebugLogLinesCounter,
+		discardedWarnLogLinesCounter:  l.discardedWarnLogLinesCounter,
+		discardedErrorLogLinesCounter: l.discardedErrorLogLinesCounter,
 	}
 }