+++ log/ratelimit_test.go
@@ -23,7 +23,7 @@ func TestRateLimitedLoggerLogs(t *testing.T) {
 	assert.Equal(t, 1, c.count)
 
 	logContains := []string{"error", "Error will be logged"}
-	c.verify(t, logContains)
+	c.assertContains(t, logContains)
 }
 
 func TestRateLimitedLoggerLimits(t *testing.T) {
@@ -34,47 +34,54 @@ func TestRateLimitedLoggerLimits(t *testing.T) {
 
 	r.Errorln("error 1 will be logged")
 	assert.Equal(t, 1, c.count)
-	c.verify(t, []string{"error", "error 1 will be logged"})
+	c.assertContains(t, []string{"error", "error 1 will be logged"})
 
 	r.Infoln("info 1 will be logged")
 	assert.Equal(t, 2, c.count)
-	c.verify(t, []string{"info", "info 1 will be logged"})
+	c.assertContains(t, []string{"info", "info 1 will be logged"})
 
 	r.Debugln("debug 1 will be discarded")
 	assert.Equal(t, 2, c.count)
+	c.assertNotContains(t, "debug 1 will be discarded")
 
 	r.Warnln("warning 1 will be discarded")
 	assert.Equal(t, 2, c.count)
+	c.assertNotContains(t, "warning 1 will be discarded")
 
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
-		# HELP rate_limit_logger_discarded_log_lines_total Total number of discarded log lines per level.
-		# TYPE rate_limit_logger_discarded_log_lines_total counter
-        rate_limit_logger_discarded_log_lines_total{level="debug"} 1
-        rate_limit_logger_discarded_log_lines_total{level="warning"} 1
+		# HELP logger_rate_limit_discarded_log_lines_total Total number of discarded log lines per level.
+		# TYPE logger_rate_limit_discarded_log_lines_total counter
+        logger_rate_limit_discarded_log_lines_total{level="info"} 0
+        logger_rate_limit_discarded_log_lines_total{level="debug"} 1
+        logger_rate_limit_discarded_log_lines_total{level="warning"} 1
+        logger_rate_limit_discarded_log_lines_total{level="error"} 0
 	`)))
 
 	// we wait 1 second, so the next group of lines can be logged
 	time.Sleep(time.Second)
 	r.Debugln("debug 2 will be logged")
 	assert.Equal(t, 3, c.count)
-	c.verify(t, []string{"debug", "debug 2 will be logged"})
+	c.assertContains(t, []string{"debug", "debug 2 will be logged"})
 
 	r.Infoln("info 2 will be logged")
 	assert.Equal(t, 4, c.count)
-	c.verify(t, []string{"info", "info 2 will be logged"})
+	c.assertContains(t, []string{"info", "info 2 will be logged"})
 
 	r.Errorln("error 2 will be discarded")
 	assert.Equal(t, 4, c.count)
+	c.assertNotContains(t, "error 2 will be discarded")
 
 	r.Warnln("warning 2 will be discarded")
 	assert.Equal(t, 4, c.count)
+	c.assertNotContains(t, "warning 2 will be discarded")
 
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
-		# HELP rate_limit_logger_discarded_log_lines_total Total number of discarded log lines per level.
-		# TYPE rate_limit_logger_discarded_log_lines_total counter
-        rate_limit_logger_discarded_log_lines_total{level="debug"} 1
-        rate_limit_logger_discarded_log_lines_total{level="error"} 1
-        rate_limit_logger_discarded_log_lines_total{level="warning"} 2
+		# HELP logger_rate_limit_discarded_log_lines_total Total number of discarded log lines per level.
+		# TYPE logger_rate_limit_discarded_log_lines_total counter
+        logger_rate_limit_discarded_log_lines_total{level="info"} 0
+        logger_rate_limit_discarded_log_lines_total{level="debug"} 1
+        logger_rate_limit_discarded_log_lines_total{level="error"} 1
+        logger_rate_limit_discarded_log_lines_total{level="warning"} 2
 	`)))
 }
 
@@ -87,19 +94,25 @@ func TestRateLimitedLoggerWithFields(t *testing.T) {
 
 	loggerWithFields.Errorln("Error will be logged")
 	assert.Equal(t, 1, c.count)
-	c.verify(t, []string{"key", "value", "error", "Error will be logged"})
+	c.assertContains(t, []string{"key", "value", "error", "Error will be logged"})
 
 	logger.Infoln("Info will not be logged")
+	c.assertNotContains(t, "Info will not be logged")
+
 	loggerWithFields.Debugln("Debug will not be logged")
+	c.assertNotContains(t, "Debug will not be logged")
+
 	loggerWithFields.Warnln("Warning will not be logged")
+	c.assertNotContains(t, "Warning will not be logged")
 	assert.Equal(t, 1, c.count)
 
 	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
-		# HELP rate_limit_logger_discarded_log_lines_total Total number of discarded log lines per level.
-		# TYPE rate_limit_logger_discarded_log_lines_total counter
-        rate_limit_logger_discarded_log_lines_total{level="info"} 1
-        rate_limit_logger_discarded_log_lines_total{level="debug"} 1
-        rate_limit_logger_discarded_log_lines_total{level="warning"} 1
+		# HELP logger_rate_limit_discarded_log_lines_total Total number of discarded log lines per level.
+		# TYPE logger_rate_limit_discarded_log_lines_total counter
+        logger_rate_limit_discarded_log_lines_total{level="info"} 1
+        logger_rate_limit_discarded_log_lines_total{level="debug"} 1
+        logger_rate_limit_discarded_log_lines_total{level="warning"} 1
+        logger_rate_limit_discarded_log_lines_total{level="error"} 0
 	`)))
 }
 
@@ -159,10 +172,14 @@ func (c *counterLogger) WithFields(fields Fields) Interface {
 	return c
 }
 
-func (c *counterLogger) verify(t *testing.T, logContains []string) {
+func (c *counterLogger) assertContains(t *testing.T, logContains []string) {
 	for _, content := range logContains {
 		require.True(t, bytes.Contains(c.buf.Bytes(), []byte(content)))
 	}
+}
+
+func (c *counterLogger) assertNotContains(t *testing.T, content string) {
+	require.False(t, bytes.Contains(c.buf.Bytes(), []byte(content)))
 }
 
 func newCounterLogger(buf *bytes.Buffer) *counterLogger {